LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/offsets-into-the-media-resource/currentTime-expected.txt


PASS currentTime initial value 
FAIL setting currentTime when readyState is HAVE_NOTHING assert_equals: expected 1.7976931348623157e+308 but got 0
PASS setting currentTime when readyState is greater than HAVE_NOTHING 
